@@ -28,64 +28,64 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^6.1.8",
-    "@angular/common": "^6.1.8",
-    "@angular/compiler": "^6.1.8",
-    "@angular/core": "^6.1.8",
-    "@angular/forms": "^6.1.8",
-    "@angular/http": "^6.1.8",
-    "@angular/platform-browser": "^6.1.8",
-    "@angular/platform-browser-dynamic": "^6.1.8",
-    "@angular/pwa": "^0.6.8",
-    "@angular/router": "^6.1.8",
-    "@angular/service-worker": "^6.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^3.2.2",
-    "@ng-select/ng-select": "^2.9.1",
-    "@ngrx/effects": "^6.1.0",
-    "@ngrx/router-store": "^6.1.0",
-    "@ngrx/store": "^6.1.0",
+    "@angular/animations": "~7.1.0",
+    "@angular/common": "~7.1.0",
+    "@angular/compiler": "~7.1.0",
+    "@angular/core": "~7.1.0",
+    "@angular/forms": "~7.1.0",
+    "@angular/http": "~7.1.0",
+    "@angular/platform-browser": "~7.1.0",
+    "@angular/platform-browser-dynamic": "~7.1.0",
+    "@angular/pwa": "^0.11.3",
+    "@angular/router": "~7.1.0",
+    "@angular/service-worker": "~7.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.1",
+    "@ng-select/ng-select": "^2.13.2",
+    "@ngrx/effects": "~7.0.0",
+    "@ngrx/router-store": "~7.0.0",
+    "@ngrx/store": "~7.0.0",
     "@types/googlemaps": "^3.30.13",
     "bootstrap": "^4.1.3",
     "core-js": "^2.5.7",
     "hamburgers": "^0.9.3",
-    "ngrx-store-localstorage": "^5.0.1",
+    "ngrx-store-localstorage": "^5.1.0",
     "rxjs": "^6.3.3",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.3",
-    "@angular-devkit/build-ng-packagr": "~0.8.3",
-    "@angular/cli": "~6.2.3",
-    "@angular/compiler-cli": "^6.1.8",
-    "@angular/language-service": "^6.1.8",
-    "@ngrx/store-devtools": "^6.1.0",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular-devkit/build-ng-packagr": "~0.11.0",
+    "@angular/cli": "~7.1.4",
+    "@angular/compiler-cli": "~7.1.0",
+    "@angular/language-service": "~7.1.0",
+    "@ngrx/store-devtools": "~7.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.4.4",
+    "codelyzer": "~4.5.0",
     "concurrently": "^4.0.1",
     "faker": "^4.1.0",
     "http-server": "^0.11.1",
     "jasmine-core": "~2.99.1",
-    "jasmine-marbles": "^0.3.1",
+    "jasmine-marbles": "^0.4.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "json-server": "^0.12.1",
-    "karma": "~1.7.1",
+    "json-server": "^0.14.0",
+    "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
+    "karma-cli": "^2.0.0",
     "karma-coverage": "^1.1.2",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-junit-reporter": "^1.2.0",
-    "ng-packagr": "^4.2.0",
-    "node-sass": "^4.9.3",
+    "ng-packagr": "^4.4.5",
+    "node-sass": "^4.11.0",
     "prettier": "1.15.3",
-    "protractor": "~5.4.1",
-    "ts-node": "~5.0.1",
-    "tsickle": "~0.32.1",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tsickle": "^0.34.0",
     "tslib": "^1.9.3",
     "tslint": "~5.11.0",
-    "typescript": "2.9.2"
+    "typescript": "~3.1.6"
   }
 }
