@@ -15,7 +15,7 @@ import {
 } from '@spartacus/core';
 
 import { Observable } from 'rxjs';
-import { tap } from 'rxjs/operators';
+import { tap, filter } from 'rxjs/operators';
 import { Card } from '../../../../ui/components/card/card.component';
 
 @Component({
@@ -47,7 +47,7 @@ export class ShippingAddressComponent implements OnInit {
 
     this.existingAddresses$ = this.userService.getAddresses().pipe(
       tap(addresses => {
-        if (this.cards.length === 0) {
+        if (this.cards.length === 0 && addresses) {
           addresses.forEach(address => {
             const card = this.getCardContent(address);
             if (
@@ -58,7 +58,8 @@ export class ShippingAddressComponent implements OnInit {
             }
           });
         }
-      })
+      }),
+      filter(Boolean)
     );
   }
 
