@@ -84,7 +84,10 @@ export class CmsService {
    * @param rootUid : the uid of the root navigation node
    * @param itemList : list of items (with id and type)
    */
-  loadNavigationItems(rootUid: string, itemList: any[]) {
+  loadNavigationItems(
+    rootUid: string,
+    itemList: { id: string; superType: string }[]
+  ) {
     this.store.dispatch(
       new fromStore.LoadNavigationItems({
         nodeId: rootUid,
