@@ -24,11 +24,13 @@ const router = {
     cmsRequired: false
   }
 };
-const mockRoutingService = {
+
+class MockRoutingService {
   getRouterState() {
     return of(router);
   }
-};
+}
+
 describe('Component Effects', () => {
   let actions$: Observable<any>;
   let service: OccCmsService;
@@ -47,7 +49,7 @@ describe('Component Effects', () => {
         { provide: CmsConfig, useValue: defaultCmsModuleConfig },
         fromEffects.ComponentEffects,
         provideMockActions(() => actions$),
-        { provide: RoutingService, useValue: mockRoutingService }
+        { provide: RoutingService, useClass: MockRoutingService }
       ]
     });
 
