@@ -1,26 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { StoreModule } from '@ngrx/store';
-import { EffectsModule } from '@ngrx/effects';
-
-import { reducerToken, reducerProvider } from './store/reducers/index';
-import { effects } from './store/effects/index';
-import { metaReducers } from './store/reducers/index';
-
-import { services } from './facade/index';
-
 import { MultiStepCheckoutModule } from './components/multi-step-checkout/multi-step-checkout.module';
 import { CartComponentModule } from './../cart/cart.module';
 
 import { guards } from './guards/index';
+import { CheckoutModule } from '@spartacus/core';
 @NgModule({
   imports: [
     CommonModule,
-    CartComponentModule,
     MultiStepCheckoutModule,
-    StoreModule.forFeature('checkout', reducerToken, { metaReducers }),
-    EffectsModule.forFeature(effects)
+    CartComponentModule,
+    CheckoutModule
   ],
-  providers: [reducerProvider, ...services, ...guards]
+  providers: [...guards]
 })
-export class CheckoutModule {}
+export class CheckoutComponentModule {}
