@@ -10,14 +10,14 @@ import {
   Address,
   PaymentDetails,
   Order,
+  CheckoutService,
+  CheckoutAddress,
   Cart
 } from '@spartacus/core';
 
 import { BehaviorSubject, Observable, of } from 'rxjs';
 
 import createSpy = jasmine.createSpy;
-
-import { CheckoutService } from './../../../facade/checkout.service';
 
 import { MultiStepCheckoutComponent } from './multi-step-checkout.component';
 
@@ -47,7 +47,7 @@ class MockCheckoutService {
   }
 }
 
-const mockAddress: Address = {
+const mockAddress: CheckoutAddress = {
   id: 'mock address id',
   firstName: 'John',
   lastName: 'Doe',
