@@ -1,10 +1,15 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { Input, Component } from '@angular/core';
 import { By } from '@angular/platform-browser';
-import { CartService, UserService, Cart, OrderEntry } from '@spartacus/core';
+import {
+  CartService,
+  UserService,
+  Cart,
+  OrderEntry,
+  CheckoutService
+} from '@spartacus/core';
 import { BehaviorSubject, of } from 'rxjs';
 import createSpy = jasmine.createSpy;
-import { CheckoutService } from '../../../facade/checkout.service';
 import { ReviewSubmitComponent } from './review-submit.component';
 
 const mockCart: Cart = {
