@@ -1,9 +1,12 @@
 import { Component } from '@angular/core';
+import { CmsComponentData } from './../../cms/components/cms-component-data';
+import { CmsBreadcrumbsComponent } from '@spartacus/core';
 
-import { AbstractCmsComponent } from '../../cms/components/abstract-cms-component';
 @Component({
   selector: 'cx-breadcrumb',
   templateUrl: './breadcrumb.component.html',
   styleUrls: ['./breadcrumb.component.scss']
 })
-export class BreadcrumbComponent extends AbstractCmsComponent {}
+export class BreadcrumbComponent {
+  constructor(public component: CmsComponentData<CmsBreadcrumbsComponent>) {}
+}
