@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 
-import { CmsPageGuards } from '../../../cms/guards';
+import { CmsPageGuards } from '../../../cms/guards/cms-page.guard';
 import { StoreFinderPageComponent } from './store-finder-page.component';
 import { StoreFinderPageLayoutModule } from '../../layout/store-finder-page-layout/store-finder-page-layout.module';
 // tslint:disable-next-line:max-line-length
